feat: allow to override some nested values in config (#374)
Currently it is not possible to override some values in the config if they are nested such as `SELFSERVICE_METHODS_OIDC_CONFIG_PROVIDERS`. It is still not possible to override some values that are nested inside `array` but this at least should allow to replace the whole array with some serialized JSON struct. It is not ideal but it should have some more options to deal with injecting secrets into Kratos et al.