fix(tracing): enable trace propagation for OTEL
The middleware already attempts to extract a parent trace from the HTTP
request. However, that does nothing unless a global OTEL propagator is
configured. This patch adds that configuration.

An alternative is to configure the propagator only for the otel Bridge:

    bridge := otelOpentracing.NewBridgeTracer()
    bridge.SetTextMapPropagator(...)

But considering that the bridge only exists due to the use of a
deprecated package it seems more appropriate to configure the global
propagator in the non-deprecated package.