feat: add more error info to traces (#833)
This is mostly for better Datadog compatibility. Datadog does not
natively implement OpenTelemetry and its conventions, so there is a
certain amount of impedance mismatching. Datadog also does not document
properly where certain error info needs to be stored within a span to
show up correctly in their UI. This is my best guess from the docs
scattered around their website.

We might have to iterate on this.